# frontend/openapi/tucker.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8181",
      "description": "Generated server url"
    }
  ],
  "paths": {
    "/api/recipes/{id}": {
      "get": {
        "tags": [
          "recipe-controller"
        ],
        "operationId": "byId",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RecipeResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "recipe-controller"
        ],
        "operationId": "update",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRecipeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/FoodResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/recipes": {
      "post": {
        "tags": [
          "recipe-controller"
        ],
        "operationId": "create_1",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRecipeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "201": {
            "description": "Created",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/FoodResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiError": {
        "required": [
          "message"
        ],
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "field": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "CreateRecipeIngredient": {
        "required": [
          "foodId",
          "grams"
        ],
        "type": "object",
        "properties": {
          "foodId": {
            "type": "integer",
            "format": "int64"
          },
          "grams": {
            "type": "number",
            "format": "double"
          }
        }
      },
      "CreateRecipeRequest": {
        "required": [
          "cookedWeightG",
          "ingredients",
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "cookedWeightG": {
            "type": "number",
            "format": "double"
          },
          "ingredients": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateRecipeIngredient"
            }
          }
        }
      },
      "FoodResponse": {
        "required": [
          "caloriesPer100g",
          "id",
          "kind",
          "name",
          "proteinPer100g",
          "tags"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "kind": {
            "type": "string",
            "enum": [
              "FOOD",
              "RECIPE"
            ]
          },
          "barcode": {
            "type": "string",
            "nullable": true
          },
          "caloriesPer100g": {
            "type": "number",
            "format": "double"
          },
          "proteinPer100g": {
            "type": "number",
            "format": "double"
          },
          "carbsPer100g": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "fatPer100g": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "cookedWeightG": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "ingredientCount": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "referenceFoodId": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "referenceFoodName": {
            "type": "string",
            "nullable": true
          },
          "tags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FoodTagResponse"
            }
          }
        }
      },
...
      "RecipeIngredientResponse": {
        "required": [
          "foodId",
          "grams",
          "name"
        ],
        "type": "object",
        "properties": {
          "foodId": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "grams": {
            "type": "number",
            "format": "double"
          }
        }
      },
      "RecipeResponse": {
        "required": [
          "cookedWeightG",
          "id",
          "ingredients",
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "cookedWeightG": {
            "type": "number",
            "format": "double"
          },
          "ingredients": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RecipeIngredientResponse"
            }
          }
        }
      },
...
      "FoodTagResponse": {
        "required": [
          "id",
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          }
        }
      },
...
    }
  }
}